feat: PaginatedListViewModel
