+++ p2p_wallet/Services/Repositories/BaseRepository/ListRepository/ItemViewModel.swift
@@ -109,4 +109,9 @@ class ItemViewModel<Repository: AnyRepository>: ObservableObject {
         error = err
         state = .error
     }
+    
+    /// Override data
+    func overrideData(by newData: ItemType?) {
+        handleNewData(newData)
+    }
 }
