+++ p2p_wallet/Services/Repositories/BaseRepository/ListRepository/ListViewModels/ListViewModel.swift
@@ -7,71 +7,6 @@ class ListViewModel<Repository: AnyListRepository>: ItemViewModel<Repository> {
 
     // MARK: - Actions
 
-    /// Erase data and reset repository to its initial state
-    override func flush() {
-        repository.paginationStrategy?.resetPagination()
-        super.flush()
-    }
-
-    /// Refresh data
-    override func refresh() {
-        // if pagination enabled, erase and reload all data
-        if let paginationStrategy = repository.paginationStrategy {
-            paginationStrategy.resetPagination()
-            super.reload()
-        }
-
-        // if pagination is not enabled, just refresh data without erasing current data
-        else {
-            super.refresh()
-        }
-    }
-
-    /// Fetch next records if pagination is enabled
-    func fetchNext() {
-        // assert pagination is enabled
-        guard repository.paginationStrategy != nil else {
-            return
-        }
-
-        // call request
-        super.request()
-    }
-    
-    /// Handle new data that just received
-    /// - Parameter newData: the new data received
-    override func handleNewData(_ newData: ItemType?) {
-        guard var data = self.data as? [Repository.ListItemType],
-            let newData = newData as? [Repository.ListItemType]
-        else { return super.handleNewData(newData) }
-        // append data that is currently not existed in current data array
-        data.append(contentsOf:
-            newData.filter { newRecord in
-                !data.contains { $0.id == newRecord.id }
-            }
-        )
-        super.handleNewData(data as? Repository.ItemType)
-    }
-
-//    func updateFirstPage(onSuccessFilterNewData: (([ItemType]) -> [ItemType])? = nil) {
-//        let originalOffset = offset
-//        offset = 0
-//
-//        task?.cancel()
-//
-//        task = Task {
-//            let onSuccess = onSuccessFilterNewData ?? {[weak self] newData in
-//                newData.filter {!(self?.data.contains($0) == true)}
-//            }
-//            var data = self.data
-//            let newData = try await self.createRequest()
-//            data = onSuccess(newData) + data
-//            self.overrideData(by: data)
-//        }
-//
-//        offset = originalOffset
-//    }
-
 //    /// Update multiple records with a closure
 //    /// - Parameter closure: updating closure
 //    func batchUpdate(closure: ([ItemType]) -> [ItemType]) {
